+++ R/read_picard.R
@@ -2,8 +2,10 @@ read_picard <- function(path,
                         verbose=TRUE){
     
     messager("Reading picard report.",v=verbose)
-    data.table::fread(path,
-                      skip = "LIBRARY",
-                      fill = TRUE,
-                      nrows = 1)
+    dat <- data.table::fread(path,
+                              skip = "LIBRARY",
+                              fill = TRUE,
+                              nrows = 1)
+    dat[,path:=path]
+    return(dat)
 }