+++ NEWS.md
@@ -29,6 +29,8 @@
     *nf-core/cutandrun* outputs files: 
     `read_picard`,`read_multiqc`,`read_bowtie`,
     `read_trimgalore`,`read_bam`,`read_peaks`
+    - Add file paths to each object.
+    - Add new arg `rbind_list`. 
 *  `rebin_peaks`/`compute_corr`:
     -Change default`bin_size` from 100 --> 5kb to improve efficiency 
     and align with other defaults of other packages (e.g `Signac`).